// src/proto/parameter_constraint_pb.ts
// @generated by protoc-gen-es v2.8.0 with parameter "target=ts,json_types=true"
// @generated from file parameter_constraint.proto (package types, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Constraint, ConstraintJson } from "./constraint_pb";
import { file_constraint } from "./constraint_pb";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file parameter_constraint.proto.
 */
export const file_parameter_constraint: GenFile = /*@__PURE__*/
  fileDesc("ChpwYXJhbWV0ZXJfY29uc3RyYWludC5wcm90bxIFdHlwZXMiVAoTUGFyYW1ldGVyQ29uc3RyYWludBIWCg5wYXJhbWV0ZXJfbmFtZRgBIAEoCRIlCgpjb25zdHJhaW50GAIgASgLMhEudHlwZXMuQ29uc3RyYWludEIjWiFnaXRodWIuY29tL3Z1bHRpc2lnL3JlY2lwZXMvdHlwZXNiBnByb3RvMw", [file_constraint]);

/**
 * @generated from message types.ParameterConstraint
 */
export type ParameterConstraint = Message<"types.ParameterConstraint"> & {
  /**
   * @generated from field: string parameter_name = 1;
   */
  parameterName: string;

  /**
   * @generated from field: types.Constraint constraint = 2;
   */
  constraint?: Constraint;
};

/**
 * @generated from message types.ParameterConstraint
 */
export type ParameterConstraintJson = {
  /**
   * @generated from field: string parameter_name = 1;
   */
  parameterName?: string;

  /**
   * @generated from field: types.Constraint constraint = 2;
   */
  constraint?: ConstraintJson;
};

/**
 * Describes the message types.ParameterConstraint.
 * Use `create(ParameterConstraintSchema)` to create a new message.
 */
export const ParameterConstraintSchema: GenMessage<ParameterConstraint, {jsonType: ParameterConstraintJson}> = /*@__PURE__*/
  messageDesc(file_parameter_constraint, 0);